refactor: updated component folder structure
