+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { ImageOverlay } from "./Image Overlay/ImageOverlay";
+export { Time } from "./Time/Time";
